pyproject.toml
[tool.poetry]
name = "henon-transcribe-2"
version = "0.1.0"
description = ""
authors = ["Graham Hukill <ghukill@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
boto3 = "^1.34.145"
ipython = "^8.26.0"
ruff = "^0.5.3"
duckdb = "^1.0.0"
pandas = "^2.2.2"
flask = "^3.0.3"
duckdb-engine = "^0.13.1"
python-slugify = "^8.0.4"
attrs = "^24.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
webapp = 'henon_transcribe_2.app:main'